chore: refactor cloud tasks to be dry
