+++ functions/package.json
@@ -25,7 +25,6 @@
   },
   "devDependencies": {
     "@firebase/app-types": "0.x",
-    "@types/google-cloud__tasks": "^0.2.4",
     "@typescript-eslint/eslint-plugin": "^5.12.0",
     "@typescript-eslint/parser": "^5.12.0",
     "eslint": "^8.9.0",
